OPENNLP-365 Re-factoring: boxing\unboxing, extra casts, shorter loops. Thanks to Aliaksandr Autayeu for providing a patch.
git-svn-id: https://svn.apache.org/repos/asf/incubator/opennlp/trunk@1199650 13f79535-47bb-0310-9956-ffa450edef68

+++ opennlp-tools/src/main/java/opennlp/tools/ngram/NGramModel.java
@@ -100,7 +100,7 @@ public class NGramModel implements Iterable<StringList>{
       return 0;
     }
 
-    return count.intValue();
+    return count;
   }
 
   /**
